// ClientApp/src/app/models/part-shortage-kit-comment.ts
export class PartShortageKitComment {
    ME_PART_NUMBER: string;
    KEYWORD_IN_KIT: string;
    ME_KIT_DESCRIPTION: string;
    STA_BUILD: string;
    STA_FOR: string;
    STA_KIT_WO_QTY: number;
    QTY_IN_KIT: number;
    STA_BUILD_TTL_QTY_REQ: number;
    STA_BUILD_QTY_REQ_OH: number;
    STA_ALL_TTL_QTY_REQ: number;
    STA_ALL_OH: number;
    DaysInQueue2: number;
    CHKKEY: string;
    COMM: string;
}